src/bridge/requirements.txt
functions-framework
requests
google-cloud-pubsub
google-cloud-storage
google-cloud-logging
